requirements_simple.txt
# Hybrid ParManus dependencies - supports both local GGUF and Ollama
pydantic~=2.10.6

# For local GGUF models (your current setup)
llama-cpp-python>=0.2.56

# For Ollama API (optional fallback)
openai~=1.66.3
httpx>=0.27.0

# Configuration and logging
tomli>=2.0.0; python_version < "3.11"
loguru~=0.7.3

# Optional: Voice interaction (uncomment if needed)
# pyttsx3~=2.90
# SpeechRecognition~=3.10.4

# Optional: Browser automation (uncomment if needed)
# playwright~=1.51.0
# selenium~=4.26.1

# Optional: Advanced features (uncomment if needed)
# fastapi~=0.115.11
# uvicorn~=0.34.0
# numpy
# pillow~=11.1.0

